+++ src/app/app.component.ts
@@ -1,7 +1,5 @@
 import { Component } from '@angular/core';
 
-declare var $:any;
-
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',