+++ src/app/sidebar/sidebar.component.ts
@@ -1,6 +1,5 @@
 import { Component, OnInit } from '@angular/core';
 
-declare var $:any;
 
 export interface RouteInfo {
     path: string;
@@ -10,14 +9,14 @@ export interface RouteInfo {
 }
 
 export const ROUTES: RouteInfo[] = [
-    { path: '/dashboard', title: 'Dashboard',  icon: 'ti-panel', class: '' },
-    { path: '/user', title: 'User Profile',  icon:'ti-user', class: '' },
-    { path: '/table', title: 'Table List',  icon:'ti-view-list-alt', class: '' },
-    { path: '/typography', title: 'Typography',  icon:'ti-text', class: '' },
-    { path: '/icons', title: 'Icons',  icon:'ti-pencil-alt2', class: '' },
-    { path: '/maps', title: 'Maps',  icon:'ti-map', class: '' },
-    { path: '/notifications', title: 'Notifications',  icon:'ti-bell', class: '' },
-    { path: '/upgrade', title: 'Upgrade to PRO',  icon:'ti-export', class: 'active-pro' },
+    { path: '/dashboard',     title: 'Dashboard',         icon:'nc-bank',       class: '' },
+    { path: '/icons',         title: 'Icons',             icon:'nc-diamond',    class: '' },
+    { path: '/maps',          title: 'Maps',              icon:'nc-pin-3',      class: '' },
+    { path: '/notifications', title: 'Notifications',     icon:'nc-bell-55',    class: '' },
+    { path: '/user',          title: 'User Profile',      icon:'nc-single-02',  class: '' },
+    { path: '/table',         title: 'Table List',        icon:'nc-tile-56',    class: '' },
+    { path: '/typography',    title: 'Typography',        icon:'nc-caps-small', class: '' },
+    { path: '/upgrade',       title: 'Upgrade to PRO',    icon:'nc-spaceship',  class: 'active-pro' },
 ];
 
 @Component({
@@ -31,11 +30,4 @@ export class SidebarComponent implements OnInit {
     ngOnInit() {
         this.menuItems = ROUTES.filter(menuItem => menuItem);
     }
-    isNotMobileMenu(){
-        if($(window).width() > 991){
-            return false;
-        }
-        return true;
-    }
-
 }